# openapi-spec.yaml
openapi: 3.0.3
info:
  title: MyLab Platform API
  description: Complete API specification for the MyLab Platform backend
  version: 1.1.0
  contact:
    name: MyLab Platform Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.mylab-platform.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      description: Get workspace projects
      tags: [Projects]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  count:
                    type: integer

    post:
      summary: Create project
      description: Create new project
      tags: [Projects]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
                  message:
                    type: string
  /projects/{id}:
    get:
      summary: Get project
      description: Get project details
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'

    put:
      summary: Update project
      description: Update project
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Project'
                  message:
                    type: string

    delete:
      summary: Delete project
      description: Delete project (admin required)
      tags: [Projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ============================================================================
  # PROJECT STAGES (Nested)
  # ============================================================================
...
  /teams/{id}/projects:
    get:
      summary: Get team projects
      description: Get all projects assigned to a team
      tags: [Team Management]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        details:
          type: object
          description: Additional error details
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        clientOrgId:
          type: string
          format: uuid
        clientOrgName:
          type: string
        externalClientName:
          type: string
        executingOrgId:
          type: string
          format: uuid
        executingOrgName:
          type: string
        status:
          type: string
          enum: [active, completed, archived]
          default: active
        workflowMode:
          type: string
          enum: [analysis_first, trial_first]
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required:
        - name
        - executingOrgId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        clientOrgId:
          type: string
          format: uuid
        externalClientName:
          type: string
          maxLength: 255
        executingOrgId:
          type: string
          format: uuid
        workflowMode:
          type: string
          enum: [analysis_first, trial_first]
      anyOf:
        - required: [clientOrgId]
        - required: [externalClientName]
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [active, completed, archived]
        workflowMode:
          type: string
          enum: [analysis_first, trial_first]

    # ============================================================================
    # STAGE SCHEMAS
    # ============================================================================
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...